% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_one_condition.R
\name{analyzeOneCondition}
\alias{analyzeOneCondition}
\title{Analyze one condition}
\usage{
analyzeOneCondition(df, numItemsInStream, paramBounds, nReplicates = 3)
}
\arguments{
\item{df}{A dataframe that must have fields targetSP and SPE}

\item{numItemsInStream}{Number of items in the stream (assumed to be the same for all trials)}

\item{paramBounds}{With fields lower and upper, each of length two indicating min and max val for efficacy,latency,precision}

\item{nReplicates}{How many times to fit the model (with different random starting points within paramBounds)}
}
\value{
efficay, latency, precision, val (negative log likelihood), and warnings (currently disabled, don't know why)
}
\description{
Analyze one condition
}
\examples{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
analyzeOneCondition(df, 16, parameterBounds(), 1)

}
