% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_with_fit.R
\name{annotate_fit}
\alias{annotate_fit}
\title{Annotate histogram with parameter vals and stats
Returns the ggplot object you passed it, but annotated.}
\usage{
annotate_fit(g, curvesDf)
}
\arguments{
\item{g}{The ggplot object to annotate}

\item{curvesDf}{A dataframe containing columns for the parameter values for each condition}
}
\description{
Annotate histogram with parameter vals and stats
Returns the ggplot object you passed it, but annotated.
}
