% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{areaUnderGaussianBin}
\alias{areaUnderGaussianBin}
\title{Calculate area under a particular domain of a Gaussian distribution}
\usage{
areaUnderGaussianBin(binStart, binWidth, latency, precision)
}
\arguments{
\item{binStart}{x value of bin start}

\item{binWidth}{bin width}

\item{latency}{mean of Gaussian}

\item{precision}{sigma of Gaussian}
}
\value{
Area under the bin for a Gaussian distribution with the parameters mean=latency, sigma=precision
}
\description{
Calculate area under a particular domain of a Gaussian distribution
}
