% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_curves_dataframe.R
\name{calc_curves_dataframe}
\alias{calc_curves_dataframe}
\title{Helper to create x,ys of fitted curves so can plot}
\usage{
calc_curves_dataframe(df, minSPE, maxSPE, numItemsInStream)
}
\arguments{
\item{df}{the data as a dataframe. Must contain SPE and targetSP}

\item{minSPE}{smallest possible serial position error}

\item{maxSPE}{largest possible serial position error}

\item{numItemsInStream}{number of items in the RSVP stream (the stimulus)}
}
\description{
To be used with ggplot. Idea is to split dataframe into conditions,
send it to calc_curves_dataframes to fit data if not already fit and either way
calculate the curves.
}
\examples{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
calc_curves_dataframe(df, -11, 11, 16)

}
