% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_curves_dataframes.R
\name{calc_curves_dataframes}
\alias{calc_curves_dataframes}
\title{Plot histogram with fitted curve}
\usage{
calc_curves_dataframes(df, minSPE, maxSPE, numItemsInStream)
}
\arguments{
\item{df}{the data as a dataframe. Must contain SPE and targetSP}
}
\description{
To be used with ggplot. Idea is to split dataframe into conditions,
send it to calc_curves_dataframes to fit data if not already fit and either way
calculate the curves.
}
\examples{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
calc_curves_dataframes(df, -11, 11, 16)

}
