% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_with_fit.R
\name{format_p}
\alias{format_p}
\title{Format a p-value appropriately.
with < precision if small enough, otherwise = p}
\usage{
format_p(p, precision = 0.001)
}
\arguments{
\item{p}{The p-value to format}

\item{precision}{Indicate precision desired, e.g. with 0.001}
}
\value{
a string
}
\description{
Format a p-value appropriately.
with < precision if small enough, otherwise = p
}
