% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_scaled.R
\name{gaussian_scaled}
\alias{gaussian_scaled}
\title{Calculate the mixture model's Gaussian, scaled to the number of datapoints in the data}
\usage{
gaussian_scaled(efficacy, latency, precision, numObservations, minSPE, maxSPE,
  grain)
}
\description{
Used in plot_hist_with_fit etc.
}
\examples{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
calc_curves_dataframe(df, -11, 11, 16)

}
