% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_scaled.R
\name{gaussian_scaled_from_df}
\alias{gaussian_scaled_from_df}
\title{Calculate fine-grained Gaussian from parameters dataframe}
\usage{
gaussian_scaled_from_df(df, minSPE, maxSPE, grain)
}
\arguments{
\item{df}{data.frame with columns (of only length 1) efficacy, latency, precision}

\item{minSPE}{smallest possible serial position error}

\item{maxSPE}{largest possible serial position error}

\item{grain}{grain, which is width of bins for which will calculate the area in the Gaussian}
}
\description{
Calculate fine-grained Gaussian from parameters dataframe
}
\examples{
 \dontrun{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
plot_hist_with_fit(df, -11, 11, df$targetSP, 16, TRUE, TRUE, TRUE)
 }

}
