% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixRSVP.R
\docType{package}
\name{mixRSVP}
\alias{mixRSVP}
\alias{mixRSVP-package}
\title{mixRSVP: A package for mixture modeling temporal errors in RSVP data}
\description{
The mixRSVP package provides
}
\section{Understanding the mixture model}{

A mixture of the guessing distribution and a Gaussian are fit, see
\href{https://alexholcombe.wordpress.com/2015/04/05/reporting-items-from-a-stream-and-mixture-modeling-to-reveal-buffering-and-a-bottleneck/}{Alex's blogpost}
To understand the relationship of latency to the exact timing of the cued letter, see
\href{https://github.com/alexholcombe/MixtureModelRSVP/blob/master/mixtureModeling/docs/understandingLatency.png}{understandingLatency}
}

