% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2E2pilot}
\alias{P2E2pilot}
\title{Dual-RSVP performance, with letters having random orientation on each frame}
\format{A data frame with 1200 rows and 23 variables:
\describe{
\item{target}{which target, the one in the left stream (1) or the right stream (2), this row pertains to}
\item{SPE}{the serial position error}
\item{allLetterOrientations}{}
\item{allRTs}{}
\item{allResponseOrientations}{}
\item{allTargetOrientations}{}
\item{condition}{}
\item{letterSeq}{}
\item{resp}{}
\item{respSP}{}
\item{subject}{}
\item{targetSP}{}
\item{trial}{}
...
}}
\source{
Alex's GoogleDrive folder, "/Users/alexh/Google\ Drive/Backwards\ paper/secondPaper/E2/Data/Raw", first 3 Ss run.
}
\usage{
P2E2pilot
}
\description{
A dataset of what letter three different undergrads reported
on each of 400 different dual-stream RSVP trials.
}
\keyword{datasets}
