% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_setup.R
\name{parameterBounds}
\alias{parameterBounds}
\title{Set parameter bounds.}
\usage{
parameterBounds()
}
\description{
Set parameter bounds.
}
\details{
Pat found bounds were needed to
prevent over-fitting to blips in the distributions. These
values are about right in most cases, but might need tweaking if
e.g. you were analysing data with an unusually high or low item rate.
}
