% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_setup.R
\name{parametersGuess}
\alias{parametersGuess}
\title{Generate guess (random starting values) for each parameter.}
\usage{
parametersGuess(lowerBounds, upperBounds)
}
\arguments{
\item{lowerBounds}{Numeric array of length 3.}

\item{upperBounds}{Numeric array of length 3.}
}
\description{
Generate guess (random starting values) for each parameter.
}
\details{
Simply generate random numbers between the corresponding lower and upper bounds, return it in guess
}
\examples{
parametersGuess(parameterBounds()$lower, parameterBounds()$upper)
}
