% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_with_fit.R
\name{plot_hist_with_fit}
\alias{plot_hist_with_fit}
\title{Plot histogram with fitted curve}
\usage{
plot_hist_with_fit(df, minSPE, maxSPE, targetSP, numItemsInStream,
  plotContinuousGaussian, annotateIt)
}
\arguments{
\item{annotateIt}{Add text to the plot indicating fit efficacy, latency, precision, and negative log likelihood}
}
\description{
Plot histogram with fitted curve
}
\examples{
 \dontrun{
df <-  subset(P2E2pilot,subject=="CB" & target==1 & condition==1)
plot_hist_with_fit(df, -11, 11, df$targetSP, 16, TRUE, TRUE)
 }

}
