% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_to_string.R
\name{warning_to_string}
\alias{warning_to_string}
\title{Calculate area under a particular domain of a Gaussian distribution}
\usage{
warning_to_string(thisWarning)
}
\value{
a string containing all warning info
}
\description{
Calculate area under a particular domain of a Gaussian distribution
}
